# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "409": {
            "description": "El correo ya está registrado"
          }
        },
        "summary": "Crear una cuenta",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credenciales incorrectas"
          }
        },
        "summary": "Iniciar sesión",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/auth/google": {
      "post": {
        "description": "Recibe el ID token que la app obtuvo de Google, lo verifica y devuelve una sesión propia. Si el correo ya tenía cuenta con contraseña, la vincula en vez de duplicarla.",
        "operationId": "AuthController_google",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleAuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "El ID token no es válido o el correo no está verificado"
          },
          "503": {
            "description": "Falta configurar GOOGLE_OAUTH_CLIENT_IDS en el servidor"
          }
        },
        "summary": "Entrar con Google",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "description": "Entrega un access token nuevo y rota el refresh token usado.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          }
        },
        "summary": "Renovar la sesión",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "description": "Sin refreshToken en el cuerpo, cierra todas las sesiones del usuario.",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cerrar sesión",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Datos del usuario autenticado",
        "tags": [
          "Autenticación"
        ]
      },
      "patch": {
        "operationId": "AuthController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar nombre o preferencia de correos",
        "tags": [
          "Autenticación"
        ]
      },
      "delete": {
        "description": "Borra el usuario con sus vehículos, documentos y dispositivos.",
        "operationId": "AuthController_deleteAccount",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar la cuenta",
        "tags": [
          "Autenticación"
        ]
      }
    },
...
    "/api/vehicles/{vehicleId}/documents": {
      "get": {
        "operationId": "DocumentsController_list",
        "parameters": [
          {
            "name": "vehicleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar los documentos de un vehículo",
        "tags": [
          "Documentos"
        ]
      },
      "post": {
        "description": "Guarda los metadatos. En la Fase 1 el archivo queda en el teléfono; cuando exista bucket se envía storageUrl.",
        "operationId": "DocumentsController_create",
        "parameters": [
          {
            "name": "vehicleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Registrar un documento",
        "tags": [
          "Documentos"
        ]
      }
    },
    "/api/documents/{id}": {
      "delete": {
        "operationId": "DocumentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar un documento",
        "tags": [
          "Documentos"
        ]
      }
    },
    "/api/plants": {
      "get": {
        "description": "Catálogo público de la Región Metropolitana. Con lat y lng el resultado viene ordenado por cercanía e incluye distanceKm.",
        "operationId": "PlantsController_search",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Texto libre: comuna, dirección o empresa.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "comuna",
            "required": false,
            "in": "query",
            "schema": {
              "example": "La Florida",
              "type": "string"
            }
          },
          {
            "name": "lat",
            "required": false,
            "in": "query",
            "description": "Latitud del usuario. Con lat y lng el listado se ordena por cercanía.",
            "schema": {
              "example": -33.4489,
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": false,
            "in": "query",
            "schema": {
              "example": -70.6693,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlantResponse"
                  }
                }
              }
            }
          }
        },
        "summary": "Buscar plantas de revisión técnica",
        "tags": [
          "Plantas PRT"
        ]
      }
    },
...
    "/api/plants/{id}": {
      "get": {
        "operationId": "PlantsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantResponse"
                }
              }
            }
          }
        },
        "summary": "Ver una planta",
        "tags": [
          "Plantas PRT"
        ]
      }
    },
    "/api/devices": {
      "post": {
        "description": "La app lo llama al iniciar sesión y cuando Expo renueva el token.",
        "operationId": "DevicesController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Registrar el token de push del teléfono",
        "tags": [
          "Dispositivos"
        ]
      },
      "get": {
        "operationId": "DevicesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar mis dispositivos registrados",
        "tags": [
          "Dispositivos"
        ]
      }
    },
    "/api/devices/{expoPushToken}": {
      "delete": {
        "operationId": "DevicesController_remove",
        "parameters": [
          {
            "name": "expoPushToken",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Dar de baja un dispositivo (al cerrar sesión)",
        "tags": [
          "Dispositivos"
        ]
      }
    },
...
  },
  "info": {
    "title": "Mi Revisión Técnica — API",
    "description": "Contrato generado desde el código. Ver README.md.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Iván Pérez"
          },
          "email": {
            "type": "string",
            "example": "ivan@ejemplo.cl"
          },
          "password": {
            "type": "string",
            "example": "unaClaveSegura",
            "minLength": 8
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "providers": {
            "type": "array",
            "description": "Formas con las que esta cuenta puede iniciar sesión.",
            "items": {
              "type": "string",
              "enum": [
                "password",
                "google"
              ]
            }
          },
          "photoUrl": {
            "type": "object",
            "nullable": true
          },
          "emailReminders": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "providers",
          "emailReminders",
          "createdAt"
        ]
      },
      "SessionResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "accessToken": {
            "type": "string",
            "description": "Token para el header Authorization: Bearer."
          },
          "refreshToken": {
            "type": "string",
            "description": "Token de larga duración para renovar la sesión."
          },
          "expiresIn": {
            "type": "number",
            "description": "Segundos de validez del access token.",
            "example": 3600
          }
        },
        "required": [
          "user",
          "accessToken",
          "refreshToken",
          "expiresIn"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ivan@ejemplo.cl"
          },
          "password": {
            "type": "string",
            "example": "unaClaveSegura"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "GoogleAuthDto": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "description": "ID token que entrega Google en el teléfono. El backend lo verifica contra los client ID configurados; la app nunca habla con Firebase."
          }
        },
        "required": [
          "idToken"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Refresh token entregado al iniciar sesión."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Iván Pérez"
          },
          "emailReminders": {
            "type": "boolean",
            "description": "Recibir avisos de vencimiento por correo."
          }
        }
      },
...
      "DocumentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "vehicleId": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "REVISION_TECNICA",
              "SOAP",
              "PERMISO_CIRCULACION",
              "OTRO"
            ]
          },
          "name": {
            "type": "string"
          },
          "mimeType": {
            "type": "object",
            "nullable": true
          },
          "size": {
            "type": "object",
            "nullable": true
          },
          "storageUrl": {
            "type": "object",
            "nullable": true
          },
          "uploadedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "vehicleId",
          "kind",
          "name",
          "uploadedAt"
        ]
      },
      "CreateDocumentDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "REVISION_TECNICA",
              "SOAP",
              "PERMISO_CIRCULACION",
              "OTRO"
            ],
            "example": "REVISION_TECNICA"
          },
          "name": {
            "type": "string",
            "example": "revision-tecnica-2026.jpg"
          },
          "mimeType": {
            "type": "string",
            "example": "image/jpeg"
          },
          "size": {
            "type": "number",
            "example": 348212,
            "description": "Tamaño en bytes."
          },
          "storageUrl": {
            "type": "string",
            "description": "URL del archivo cuando exista almacenamiento en la nube. En la Fase 1 va vacío."
          }
        },
        "required": [
          "kind",
          "name"
        ]
      },
      "PlantResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "prt-25"
          },
          "company": {
            "type": "string",
            "example": "TÜV Rheinland Andino"
          },
          "comuna": {
            "type": "string",
            "example": "La Florida"
          },
          "address": {
            "type": "string",
            "example": "Av. Vicuña Mackenna Poniente 7955"
          },
          "lat": {
            "type": "number",
            "example": -33.515243
          },
          "lng": {
            "type": "number",
            "example": -70.60707
          },
          "classes": {
            "example": [
              "A",
              "B"
            ],
            "description": "A = livianos, B = pesados.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "phone": {
            "type": "object",
            "nullable": true
          },
          "schedule": {
            "type": "object",
            "nullable": true
          },
          "distanceKm": {
            "type": "number",
            "example": 2.4,
            "description": "Distancia en km. Solo cuando se envían lat y lng."
          }
        },
        "required": [
          "id",
          "company",
          "comuna",
          "address",
          "lat",
          "lng",
          "classes"
        ]
      },
      "RegisterDeviceDto": {
        "type": "object",
        "properties": {
          "expoPushToken": {
            "type": "string",
            "example": "ExponentPushToken[xxxxxxxxxxxxxxxxxxxxxx]",
            "description": "Token entregado por expo-notifications en el dispositivo."
          },
          "platform": {
            "type": "string",
            "enum": [
              "android",
              "ios"
            ]
          }
        },
        "required": [
          "expoPushToken"
        ]
      },
      "DeviceResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "expoPushToken": {
            "type": "string"
          },
          "platform": {
            "type": "object",
            "nullable": true
          },
          "lastSeenAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "expoPushToken",
          "lastSeenAt"
        ]
      }
    }
  }
}